embedding-worker/requirements.txt
sentence-transformers==2.2.2
qdrant-client==1.6.9
watchdog==3.0.0
fastapi>=0.68.0,<1.0.0
uvicorn>=0.15.0,<1.0.0
python-multipart>=0.0.5,<1.0.0
qdrant-client>=1.1.0,<2.0.0
sentence-transformers>=2.2.0,<3.0.0
watchdog>=2.1.6,<3.0.0
pydantic>=1.8.0,<2.0.0
python-dotenv>=0.19.0,<1.0.0
numpy>=1.21.0,<2.0.0
tqdm>=4.62.0,<5.0.0

# Document processing
pypdf>=3.0.0,<4.0.0
python-docx>=0.8.11,<1.0.0
python-magic>=0.4.24,<1.0.0
python-magic-bin>=0.4.14,<1.0.0; platform_system=="Windows"

# Text processing
spacy>=3.5.0,<4.0.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
langchain>=0.0.200,<0.1.0
scikit-learn
scipy
nltk
sentencepiece
python-docx>=0.8.11
python-magic>=0.4.27
python-magic-bin>=0.4.14; platform_system=="Windows"

# Text processing and chunking
langchain>=0.0.200
spacy>=3.5.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl

# Utilities
tqdm>=4.65.0